backend/requirements.txt
django-filter
djangorestframework-gis
djangorestframework
django
psycopg2-binary